Split Piece class into Piece and PlayerPiece
The Piece contains only the type and color while the PlayerPiece
contains the piece and other info such as position and previous moves.

The ChessBoard will contains Piece array and the Player will has
PlayerPiece ArrayList of his active pieces.

Now it is the Player responsibility to update the ChessBoard pieces
when adding,moving and removing a piece of him.

Fix ChessBoard null pointer in Player.