Replace the PlayerPiece member in Piece class
Replace the PlayerPiece member in Piece class with the position of
the piece on the board.

Update the legal moves for all the pieces of the player after each move
and store them in a hash map and use it to check if a piece can move to
a certain square.

Show option to select the type of game between computer and human or
computer and computer in the starting activity.